+++ modules/utils/index.js
@@ -1,3 +1,4 @@
+exports.addRoutingMethods                         = require('./addRoutingMethods');
 exports.bufferStream                              = require('./bufferStream'),
 exports.compileRoute                              = require('./compileRoute'),
 exports.decodeBase64                              = require('./decodeBase64'),