+++ CHANGES.md
@@ -1,3 +1,7 @@
+### HEAD
+
+  * mach.stack respects ordering of routes as well as mappings
+
 ### 0.11.0 / 2014-07-11
 
   * Add route parameters to request.params instead of passing them